Module 5 - RESTful Web API

+++ CartService/CartService/Application/UseCases/CartItems/Commands/RemoveItemFromCartCommand.cs
@@ -1,4 +1,5 @@
-﻿using CartService.Application.Interfaces;
+﻿using Ardalis.GuardClauses;
+using CartService.Application.Interfaces;
 using MediatR;
 
 namespace CartService.Application.UseCases.CartItems.Commands
@@ -21,6 +22,15 @@ namespace CartService.Application.UseCases.CartItems.Commands
 
 		public async Task Handle(RemoveItemFromCartCommand request, CancellationToken cancellationToken)
 		{
+			var cartEntity = await _repository.GetCartItems(request.CartId);
+			if (cartEntity == null || !cartEntity.Any())
+				throw new NotFoundException(request.CartId.ToString(), nameof(cartEntity));
+
+			var cartItemEntity = cartEntity
+				.Where(item => item.Id == request.ItemId)
+				.FirstOrDefault();
+			Guard.Against.NotFound(request.ItemId, cartItemEntity);
+
 			await _repository.RemoveItemFromCart(request.CartId, request.ItemId);
 		}
 	}
